requirements.txt
Adafruit-Blinka==5.9.2
adafruit-circuitpython-ads1x15==2.2.4
adafruit-circuitpython-busdevice==5.0.1
Adafruit-PlatformDetect==2.23.0
Adafruit-PureIO==1.1.8
aiohttp==3.5.4
aiohttp-jinja2==1.2.0
aioopenssl==0.4.1
aiosasl==0.4.0
aioxmpp==0.10.5
arrow==0.15.5
async-timeout==3.0.1
asynctest==0.13.0
attrs==19.3.0
Babel==2.8.0
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Adafruit_DHT
dnspython==1.16.0
factory-boy==2.12.0
Faker==4.0.2
gas-detection==1.0.1
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.5.0
Jinja2==2.10.3
jinja2-time==0.2.0
lxml
MarkupSafe==1.1.1
more-itertools==8.2.0
multidict==4.5.2
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
psutil==5.8.0
py==1.8.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyftdi==0.52.0
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyserial==3.5
pytest==5.4.1
pytest-asyncio==0.10.0
python-dateutil==2.8.1
pytz==2019.3
pyusb==1.1.0
RPi.GPIO==0.7.0
singletonify==0.2.3
six==1.14.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
spade==3.1.4
text-unidecode==1.3
timeago==1.0.8
typing-extensions==3.7.4.3
tzlocal==1.5.1
wcwidth==0.1.8
yarl==1.4.2
zipp==3.1.0
